refactor: move 'BaseController' inside 'Core'
